Add async context managers for watches
Testing of watches is hard because the information of a watch was
created is not propagated properly. This leads to situations where a
resource is modified before the watch is established leading to
non-deterministic testing behavior.

Hence, context managers are introduced ensuring that a watch is
created on entering. This means a user of this new API can assume that
a watch is established inside these watch contexts.