Register Kubernetes services endpoints
Extend ApplicationStatus with the "services" dictionary, associating
each Service to its endpoint. Implement a simple event dispatcher for
registering arbitrary listeners for events. Use this event dispatching
mechanism to regster a listener storing the service endpoints in the
Kubernetes application resource. If there are further tasks for
specific Kubernetes resources, the event dispatching mechanism can be
used for these use cases as well.

Do not create a service in the test_app_creation unit test for two
reasons:

1. Simplification: no need to mock the Kubernetes Service API
2. Scope: app creation is tested, not registration of service endpoints