Refactor MagnumSpec to have composition
First step for the issue #234

Instead of having a kind of ClusterSpec for each type of k8s Cluster,
we will have a separate resource for the k8s cluster providers (e. g.
Magnum), or no resource in the case of an external cluster.

A future "MagnumPlatform" resource would be used to create a
KubernetesCluster Resource, which would be handled independently.