Add Cluster deletion while protecting Applications
Added after modifications of issue #234, but following
request from issue #227.

While deleting a Cluster, we want to ensure that the
administrator is aware of Applications that may depends on the
Cluster, thus the API prevents him/her from doing it.
If the Administrator still wants to delete the cluster, he/she needs
to use the --cascade argument.