Use finalizers to cleanup Kubernetes applications and clusters
This is a step in the direction of garbage collection (see #235).

If a Kubernetes application is created or updated by a worker, a
"cleanup" finalizer is added preventing the API from deleting the
resource immediately. The application transitions into the "deleting"
state on a DELETE request instead (aka. setting the ".metadata.deleted"
attribute). If a resource is in the "deletion in progress" / "deleting"
state finalizer can only be removed. Then, the Kubernetes application
controller handles the cleanup of all spawned Kubernetes resources
and removes the "cleanup" finalizer. When removing the last finalizer,
the API deletes the Kubernetes application resource from the database.
This will be handled by the garbage collector in the future.

If a Kubernetes cluster is deleted but there are applications depending
on this cluster, the cluster transitions into the "deleting" state and
a "cascading_deletion" finalizer is added. The Kubernetes cluster
controller handles the "cascading_deletion" finalizer by issuing the
deletion of all depending Kubernetes application and waits for the
deletion to finish. After this, the "cascading_deletion" finalizer is
removed. When removing this last finalizer, the API deletes the
Kubernetes cluster from the database.