Add modifications for Ansible context
See discussions on issue #240

The 'enabled' field is added to the TLS configuration of rok and the
krake controllers to allow easy settings with Ansible.

The scheme ("http" or "https") is removed from configuration of krake
controllers and rok, it is instead deduced from the rest of the
settings. This again simplify the settings