Add garbage collector as controller
See #235

The Garbage Collector was added to allow deletion of resources NOT to be
done on the API. Instead, the cleanup is only done by the controllers.

The Garbage Collector mark as deleted all dependents of a resource marked
as deleted. It removes the "cascading_dependencies" finalizer, and
delete all resources without finalizer.

This first iteration is done with the GC being a Controller with many
workarounds.