ansible: Add option to deploy VMs using Config Drive
The metadata service can be unstable and the VMs unable to reach
it, download their metadata (especially SSH authorized keys), and
therefore unreachable.

In order to workaround this issue, VMs can be configured to use a
Config Drive to access their metadata instead of connecting to the
metadata server.

This commit adds this option, defined in `group_vars.all.yaml` and
passes it up to the stack for Minikube, Krake, and the Gateway VM.