Improve automation of garbage collection
See issue #235

The first iteration of the Garbage Collector had all resources it needed
to handle hardcoded as a list. This is obviously not acceptable for
extensibility issues. The resources are now added via a decorator.

The API clients can also now be easily extended, but still need to be
given to the Garbage Collector as list. Because these clients need to be
imported to be generated, not using the list but a decoration mechanism
means we would anyway be forced to import them. The decorator for this
would need to be on another module than garbage_collection to prevent
circular imports. We would then be forced to add imports needed for Garbage
Collection on ANOTHER module, which is not explicit enough.