WORKAROUND Fix deletion failed apps not supported
The MR 237 should not have been merged for the moment, as it uses some
changes done on the declarative API branch, not merged yet.

Applications with the FAILED state could not be deleted. The presence
of the FAILED state of an Application is normally used to know to force
delete it. But with the current state of master, if you delete a FAILED
application, it changes into a DELETING state, and the Reason is
removed.

So the workaround keeps the Reason, and if present, it means the
Application could not be deleted and needs to be force deleted.