Remove calls to API, keep only database calls
See issue #235

The Garbage Collector used a mix of API calls and database calls through
database sessions. This was not optimal. Also, the API Client objects
needed to be hardcoded to be registered.

The GC is a Controller, and it is preferable that a Controller only
communicates with the API through a client.

However, to delete an element definitely, one needs to communicate with the
database, as we cannot allow such call on the API. As such, it was
decided to only use database calls. This means, the GC is the only
Controller with direct access to the database, and can only access it,
not the API through any client.

Thus, all API calls have been changed into database calls.