Implement reconciliation for Kubernetes application
All controller operations in the Kubernetes cluster are idempotent.
This removes the requirement for atomic operations where every
change on the Kubernetes cluster needs to be reflected in the
application status fields. Instead, all operations are performed
sequentially and the status field is updated only one time. If the
controller crashes inbetween two operations for some reason, the
controller can just replay all operations.

Rename "resources" status field into "manifest" for consistency with
the application specification.