Add debouncing on the WorkerQueue
See issue #235

To prevent too many updates at once on a Controller, the WorkerQueue has
been modified to allow debouncing. When an update is received, it is not
directly given to the Workers, but with a timeout. If a new update on
the resource is available, the timeout is reset. This mechanism works
because the system is driven by the state of the resources. If a
resource is updated two times without much delay, only the last status
should matter.