Add worker for testing resource reception based on UIDs
A resource can be received multuple times during a watch-and-list call
of a controller when the watcher sends the same resource that is
already returned by the list request. This means the precise number of
received resources cannot be used in assertions. Instead, the new
worker validates received resource's UIDs against a set of expected
UIDs.

The worker can be reused in multiple tests. Hence, it was implemented
in the "krake/tests/controller/__init__.py" module.

Fixes #252