Remove handling of finalizers by garbage collector
See issue #235

The garbage collector used to have a "cascading_deletion" finalizer,
which was blocking the deletion of a resource as long as it had
dependents.

The garbage collector now simply checks if dependents are still present,
if yes, it does not delete the resource.

Also, the "cleanup" finalizer was added by the API, which is not
optimal, as the finalizer should be added and removed by the same
component, in this case, the controller of the resource. This finalizer
can now only be added by the Controller.