Use patch instead of replace operation to update Kubernetes resources
Kubernetes replace operation updates every field of a resource. If a
field is missing, Kubernetes assumes that this field should be cleared.
The user manifest is only a partial representation of the Kubernetes
resource. Hence, replacing the existing Kubernetes resource with this
partial representation leads to errors because immutable fields get
potentially cleared out, e.g. "spec.clusterIP".

There are two ways of solving this issue: Either read the Kubernetes
resource first, update the returned object with the fields of the user
manifest and then replace the Kubernetes resource or use a "patch"
operation instead of "replace". For now, patch seems to be the easier
solution.