Handle traceback for HTTP errors in Rok
See issue #256

When the API sends a response with an error code >= 400, the response
was raised and displayed a traceback.

This commit adds a simple error handler to display just the error
message and quit. It also removes the error handling from the printing
function, which remove a side effect.