Refactor elements on Garbage Collector
See issue #235 and discussions of MR !246

The finalizer `cleanup` has been renamed to have a clear separation
between the cleanups of all kind of resources.

The tag `@see` is not conform with sphinx and has been changed.

The test for GC is moved to the controller directory, just like the
actual GC source code.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>