Use transactions for etcd operations
The database API was changed such that the etcd revision of an object
stored in etcd is now attached to the object itself. The attached
revision is used to derive transaction guards -- conditions that must
be true for a transaction to succeed.

Object without an attached revision are considered "fresh". It is
assumed that the associated etcd key of a fresh object does not exist
in etcd.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>