Clear debounce timers in task done callback
The debounce timer was previouly cleared in the "get" method of the
queue. It was assumed that the timer for fetched key is always done
after it was fetched from the queue. If a debounce timer was still
pending -- which means the key was not yet put into the queue -- and the
previous version of the key was fetched from the queue this assumption
was no longer valid.

As a solution, the key is removed in a "done callback" of the debounce
timer. Add a unit test for mentioned case of a pending debounce timer.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>