Add finalizer before creating Kubernetes objects
If the Kubernetes application controller is still in the process of
creating Kubernetes resources and adds the finalizer after all
resources are created, this leads to a problem with the garbage
collector: if the application is deleted while the application
controller is creating the Kubernetes objects, the garbage collector
does not see any finializer and removes the application from the store.
This means the controller cannot write the result back the API. What is
even worse is that the created Kubernetes objects will never be cleaned
up which is effectively a resource leakage.

This problem is prevented when adding the cleanup finalizer *before*
starting to create any real world resource. The garbage collector will
see the finalizer and wait for it to be removed. The application
controller can finish its creation work and will handle the deletion
of the application afterwards.

Reduce verbosity of info-level logging output.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>