Refactor the base Controller
See issue #259

Several functionalities were not possible for the Controller, for
instance, for the workers to have access to the WorkQueue, or having
several WorkQueues. Also the connection to the API was handled by the
Controller, which gave it too many responsibilities.

This commit adds flexibility and allows a developer to create a
Controller with different simple bricks, by simply rewriting the
__init__ or other function depending on her/his needs. Several
functionalities have also been removed from the Controller management,
such as API calls or starting the process. This is a step towards
separation of concerns.

A first iteration of Observers has been added to watch the real world
status of resources. One Observer is meant to observe one instance of
resource.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>