Rearrange pipeline stages
We previously always had two Pipelines running on MR. This commit ensures that
only one Pipeline is ran on Merge Requests. We have to define the `test` stage
twice: one to run on each commits and branches (default values) but to exclude
from merge_requests and schedules, and one specifically for merge_requests and
schedules. We use YAML anchors to avoid code duplication.

Signed-off-by: Matthias Goerens <matthias.goerens@cloudandheat.com>