Use aiohttp test server for Prometheus exporter fixture
Remove the deprecated usage of "app.loop" in the metrics exporter.

The default "prometheus_client.Gauge" uses a global collector registry
by default. A metric is only allowed to be registered once. But if the
"prometheus_exporter" fixtures is used multiple times, a new gauge
object is created every time leading to a registry error. The solution
for this problem is to create a new registry instance for each
"prometheus_exporter" fixture.

The default "prometheus_async.aio.web.server_stats()" handler uses the
global registry as well. Therefore, we create our own metrics handler
with the locally scoped registry. All related handler functions for the
Prometheus exporter aiohttp application are using function closure now.
This increases the encapsulation of the fixture.

Furthermore, the data directory of the Prometheus server is created in
the temporary directory by adding a "--storage.tsdb.path" command line
option.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>