Fix GC wrongly used session attribute.
The GC used a session attribute to store an opened database Session,
with the goal of accessing it in all methods. This worked before, when
the Controller had several workers, which were the one having the
session attribute. With the new Controller API, the GC workers are now
methods, and shared the same session. Some worker then closed the
session before it could be used by the other.

This has been fixed: a session variable is simply created when a
resource is received, and deleted when the worker did the job on the
resource.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>