Undo f2946a6 and try all fd to get the terminal size
We don't want to use shutils.get_terminal_size() as it breaks when piping. We
come back to the previous method, which was to use os.get_terminal_size().

However, in order to cover the multiple cases when stdin is not an interactive
terminal, we add some retry logic to test stdout and if necessary stderr as
well. If none of these are interactive terminals, we fallback on defaults
values.

Signed-off-by: Matthias Goerens <matthias.goerens@cloudandheat.com>