Add Response class to contain all response's information
The Response class contains all information about the command's response, such
as the output and the return code. The json representation is now generated
only once, on first need.

This also allows simplification for testing the command's return code. The run
function's signature is simplified, and the condition are now expected to
raise AssertionError when not met.

Signed-off-by: Matthias Goerens <matthias.goerens@cloudandheat.com>