Move configuration from docker-compose to volume
See issue #270

The .env file used by docker-compose was previously used to set all
variables that would configure Krake. Several environment variables were
present, and new one needed to be added for each new piece of
configuration. To automatically create this file, another layer was
needed, with a jinja2 template that needs to be manually filled.

To simplify this, the configuration files used by the Krake component
are now generated using variables in the hostvars. These files can then
easily be added to a bind mount to all Krake containers.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>