Make source files positional and destination optional
The source files argument consumes all non-option strings because of
its nargs "+" option. This means if the destination directory is given
at the end the command line it will be interpreted as source file
as well. This problem can be solved by specifying the destination
directory before the "--src-files" option which is unusal for *nix
command line tools.

Resolve this issue by making "--src-files" positional. In this way, the
files are expected to be given as trailing arguments. The destination
directory is now optional "--dst" and uses the current working
directory as default.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>