Remove environment variable as config options
See discussions on issue #266

As the configuration files have been separated and the environment
variables are not used anymore by our Docker compose, the support of the
environment variables as configuration options is dropped. Thus the
functions allowing this are simply removed.

Instead, a feature will be added later on to add automatically
command-line options for each configuration part.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>