Change the configuration dict into a Serializable
See issue #266

The configuration was previously a simple dict created from the YAML
configuration. It meant that this dict was passed to all component,
without any kind of validation on the value, nor on the structure of the
dictionary. Then, issues could occur in functions deeper in the code,
that use specific parts of the configuration.

Now, an ApiConfiguration object is used, which is still created from the
dictionary obtained by the YAML file, but is a Serializable. As such,
the configuration's structure and values are validated by Marshmallow.
If there is a misformed configuration for a component, an exception is
raised and the component cannot start.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>