Reschedule Kubernetes applications on spec changes
Additional status information is required for properly reschedule
Kubernetes Applications. The following fields are added to the
status:

```yaml
status:
  scheduled:
  scheduled_to:
  running_on:
```

The states of Kubernetes Applications are reworked as follows:

- PENDING (initial state)
- CREATING (first time Kubernetes objects are created)
- RECONCILING (spec changed and Kubernetes objects are updating)
- MIGRATING (cluster changed)
- DELETING (Kubernetes objects are deleting)
- FAILED (error occurred)

Every state transition is performed by the controller. The API does
not change the state any longer.

Migration of Kubernetes Applications is implemented in the controller.
For now, the scheduler simply selects the best matching cluster based
on the weighted sum of all metrics. A migration is initiated if the
"scheduled_to" field points to a different cluster than the
"running_on" field. The migration is performed in a very naive form:
Kubernetes objects are deleted on the old cluster ("running_on"). After
that they are recreated on the new cluster ("scheduled_to").

Use HTTP endpoint mockups for Prometheus in controller unittests to
increase the speed of pytest runs. The correctness of the Prometheus
API mockup is ensured in another unittests. This decouples the
testing of controllers functionality from Prometheus.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>