Refactor Controller client registration
The name of the method "create_background_task" of the Controller was
misleading, as it suggested that the Controller background tasks were
created there. However, the method was only there to create and
prepare the objects or coroutines that would be used later on as tasks.
Thus, it has been renamed to correspond to its actual behavior.

The "clean_background_tasks" method suffered from the same issue as the
previous and has been changed as well

The client is also now given to the new "prepare" method, and not
expected to be set on the Controller first. This simplifies the tests
also.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>