Add dependency graph implementation
See issue #261

The first version of the garbage collector had direct access to the
database to delete and get the dependencies of resources directly. As we
want to remove this access and only let it connect to the API, the gc
needs to know which resources depends on a resource to delete.

Thus a graph implementation is added, that will be later on used by the
garbage collector to store on cache the dependencies of all resources it
watches. The garbage collector will in the end be able to get the
dependencies of any resource directly without having to ask the API.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>