Implement flexible MTU configuration for docker
This commits corrects the path of the configuration file for the docker
daemon.

This file (/etc/docker/daemon.json) configures only the docker0 bridge.
Additional "user-defined" bridges are not affected by this configuration file.

By default, and if no networks are defined in docker-compose.yml,
docker-compose automatically creates a new bridge network to run the
application. That's why the MTU value was always the default 1500, and the
application couldn't always successfully communicate with the external world.

We need to configure this "user-defined" network created by docker compose
using the "com.docker.network.driver.mtu" driver option.

We want to configure this dynamically from Ansible, as the MTU value can
differ depending on the infrastructure. We add Ansible templates.

To avoid having an Jinja2 template for the whole docker-compose file, which
would make not easily reusable outside of Ansible, we use a
docker-compose.override file which merges with the standard docker-compose
file automatically.

Additionally, https://gitlab.com/rak-n-rok/krake/issues/272#note_234983673
provides some information on the problem resolution

Signed-off-by: Matthias Goerens <matthias.goerens@cloudandheat.com>