Refactor the gc to remove the database session
The first version of the garbage collector used a direct connection to
the database, making it different than the other controller, and also
bypassing the API when processing the data stored. It could have brought
inconsistencies and made the garbage collector a very special
controller, quite different from the others.

The new garbage collector now uses only API calls to handle the
cascading deletion of resources and its direct and indirect dependents.
Thus the API also needed to be modified to allow the new behavior. The
latter is now more consistent, because the actual deletion of data is
now done by the API. It also keeps the idea of the API being the only
component that can access and process the data.

See issue #261

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>