Add mockup server for Prometheus query HTTP API
The little script is built on top of the existing "make_prometheus"
function used already in the scheduler unit tests.

The metric defintions are provided via command line options. An
additional special HTTP endpoint "/-/update" is provided that can
be used to update the metric defintions served by the mockup server.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>