Fix minikube provisionning due to version change
The pipeline previously failed with the newest version release of
minikube, for which the output of the status command slighly changed.

To prevent such an issue in the future, the versions of minikube and
kubectl have been pinpointed. Later on, the versions will be actively
updated when decided, instead of passively receiving errors on update.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>