Add simple "stickiness" to Kubernetes applications
Previously, the scheduling rank of a Kubernetes cluster was independent
of the currently selected cluster. This leads to a "flip-flop" behavior
where applications are scheduled between clusters with nearly
equivalent ranks.

"Stickiness" is introduced to circumvent this problem: a virtual weight
is added to clusters based on the current location of the application.
The rank of the cluster where the application is currently scheduled to
is increased. This makes the application "stick" to its currently
scheduled cluster. The difference in rank must be larger than the given
stickiness weight.

The optimization strategy of the scheduler was flipped to make this
mechanism work: the scheduler searches for clusters with MAXIMAL rank.
If there are multiple maximal clusters, randomly select one of them.
Because we are using floating point numbers for ranks this situation
should be very rare in practice.

For now, this stickiness is just a simple factor that is statically
configured for the scheduler but it easy to extend this mechanism with
more advanced decisions, e.g. a "aging" approach where the stickiness
dynamically decreases over time.

The location of metric weights was moved from the metric itself to the
cluster. This allows the same metric to have different weights for
different clusters. This might be desirable because a metric can have
different roles or importance across clusters.

Furthermore, cluster ranks are now completly normalized: a cluster rank
is now guaranteed to be in range [0; 1]. The value of each metric is
same is normalized as well as the sum of all weights.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>