Add ".update()" method to serializable objects
The previous "copy_fields()" function did not play well with
polymorphic containers because of its dynamic fields. Hence, the
responsiblity of updating a serializable data class is moved to the
object itself.

The in-place modification via ".update()" was chosen over a copying
approach via ".merge()" which would return a new object with the merge
result. The main reason is consistency: database operations are also
modifying a data class by assigning the revision to the object. This
metadata needs to be copied as well but a serializable object should
be independent of etcd-specific code. A new function
"krake.api.database.merge()" is needed for handling the database-
specific attributes. But the problem remains that we follow two
different principles here: database operations are modifying objects
in-place while ".merge()" returns a new object. Returning a new object
has the benefit of avoiding side-effects. The database operations could
be rewritten in order to keep the data class untouched and return
metadata as an extra object. We had this implementation already in the
past but we moved away because it was annoying to handle the revision
metadata everywhere. It is more elegant to encapsulate the whole state
-- including database metadata like the revision -- inside one object.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>