Search for configuration files in extra function
The signature of "load_config()" was confusing because the function
basically implements an if-then-else logic of a configuration file
was specified via command line or not. The responsibility of searching
and loading is not split into two different functions which makes the
code easier to understand and the control more explicit.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>