Move krake configuration from host to group
To configure the Krake installed by Ansible, files are generated by the
configure/generate script. The values given were previously taken from
the hosts file. This led to the issue that the hosts for the local tests
and for the CI need to hold the configuration. In the pipeline, the CI
hosts did not have this configuration (only the local hosts), and Krake
was misconfigured, as the generate script had to use default values.

Now, the configuration is moved to the group vars of Krake. This allows
a central configuration, that is then used for the CI pipeline and the
local tests.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>