Encapsulate option value replacement into a mapper
As the configuration read from file can now be overridden by the
configuration given by command line, these options need to be read and
used to replace the right element of the configuration, and recursively.

This was previously done by functions that called each other, and some
variables needed to be passed from one to the other. Now, a special
mapping class is handling all these calls, and only one instance of this
class is needed per Krake component. All complex operations are now
hidden behind this mapper.

See discussion on MR !265

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>