Allow receiving whole serialized data on the API
Previously, the API handlers, when receiving a resource, had access only
to the fields that are not readonly. This made sense, as the API does
not need the uid of a resource that is already in the database for
instance, because the uid will not be modified again.

However, in some cases, information such as the modified timestamp can
differ between what a Controller sent through an update and what is
stored in the API (as an example). It also makes sense to send more
information to the API, even if this information may not be used.

Thus, the API now expects the whole resource to be sent to it, when not
otherwise requested. However, when sending a create request, few
elements (such as uid, created timestamp...) are to be set by the API,
not the client. These fields are read-only. In such cases, these fields
are excluded from the expected fields. It means the client can send
incomplete resources. In other cases, the API will send a 422 code
("Unprocessable Entity") if fields are not valid.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>