Create Reason and Reason codes for HTTP on the API
On the API, when errors occurred, it would send an HTTP answer with a
status code above 400 and a reason message. These messages have no
structure, and are simply strings to inform the client.

To support some kind of comparison between the errors, a reason code is
introduced for the HTTP communications between the components. This code
is encapsulated in an HttpReason object to store the message, for
debugging and readability.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>