Add "__post_init__()" mechanism to Serializable
The main motivation for this addition is an easy intergration of
schema-level validation. Instead of supporting marshmallow's validation
decorators on the Serializable-level, e.g.

```python
class Interval(Serializable):
    min: int
    max: int

    @validates_schema
    def _validate(cls, data, *, partial, many):
        if data["min"] > data["max":
            raise ValidationError(
                "'min' must not be greater than 'max'."
            )
```

we leverage the standard dataclass mechanism of "__post_init__()" to do
the validation. The above example becomes simply:

```python
class Interval(Serializable):
    min: int
    max: int

    def __post_init__(self):
        if self.min > self.max:
            raise ValidationError(
                "'min' must not be greater than 'max'."
            )
```

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>