Add Observer for the Kubernetes Applications
When Applications are created by the Kubernetes Controller, actual
Kubernetes resources are created on clusters. These resources may then
be modified independently from Krake (deletion, update, or cluster not
reachable anymore). Krake needs to be aware of these changes.

To achieve this, the Observers introduced with the Controller
refactoring are used. A Kubernetes observer is implemented to watch the
current status of all resources managed by Krake on Kubernetes clusters.

See issue #278

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>