Fix dependents wrongly put in queue
When deleting a dependent, its dependencies were put into the queue even
if they were not supposed to be deleted. This triggered the deletion of
all other dependents. For instance, if A is the owner of B and C, and B
was deleted, A would be put into the garbage collector queue. It would
not be deleted, but C would be marked for deletion.

This is now fixed by simply not putting dependencies in the queue if
they are not marked for deletion.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>