Ensure key is enqueued only once in WorkQueue
Multiple subsequent puts in the WorkQueue without consuming in-between
leads to enqueing of the same value multiple times into the internal
asyncio queue because the key is not added to the "processing" set yet.
This leads to a violation of the WorkQueue guarantee.

This bug is fixed by adding the key into the "processing" set when the
key is put into the internal queue. This ensures that the key is only
enqueued once in the internal queue. The "processing" was renamed into
"active" to reflect that it ensures uniqueness accross the currently
processed and enqueued keys.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>