Add Magnum controller
The controller is responsible of creating a the Magnum cluster in
OpenStack. This is done by the official "magnumclient". Unfortunately,
the client is synchronous. Therefore, every OpenStack API call is
performed in a thread executor to not block the asyncio event loop.

Create Kubernetes cluster ontop of Magnum cluster. Use simple finite
state machine for reconcilation. This reduces redundant code for
handling state transitions.

Check access during reconcilation of Magnum clusters. If a client error
occurs, the kubeconfig of the associated Kubernetes cluster resource
is updated.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>