Fix infinite client read timeout
Client timeouts in aiohttp are expressed as floats but Darwin's kqueue
selector does support integers (or None) as timeout parameter only.

We used the special float("inf") value to express an never expiring
read timemout. The correct way of specifying an infinite timeout is
using "None". This resolves to problem with the kqueue selector as
well.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>