Fix check for generic types
Types from the "typing" module should not be used in "issubclass()" or
"isinstance()" checks. The way to go for checking these types is
to leverage the information of the "original" type.

Version-specific implementations are used for checking generic types
because the internal implementation of the typing module is not stable
yet.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>