Do not catch exceptions during etcd version retrieval
The version of the etcd cluster and server is resolved synchronously
(using the "requests" library) in "__init__" of the client. The
original implementation swallos every exception which potentially
leads to a state where required attributes are not set leading to
attribute errors.

The patched implementation does not catch any exception.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>