Add scheduling support for MagnumCluster
OpenStack Magnum clusters follow the same principles as Kubernetes
applications: there are label constraints for the underlying OpenStack
project and projects are ranked based on metrics.

For now, labels and metric references of the underlying OpenStack
projects are copied to the Magnum cluster resources. The concept here
is that resources built on-top of other resources "inherit" their
labels and metrics. The copying approch is not optimal because it
introduces redundancy where the resource states can run out-of-syn. A
better approach would be adding support for transitive label and metric
lookups in the scheduler.

There is currently no support in the scheduler for spawning a Magnum
cluster adhoc to fulfill the constraints of an application. It is
non-trivial to make this decision. There is no consensus how this
problem should be solved.

Adjust rok to support the new OpenStack API resource interface.

Signed-off-by: Lucas Kahlert <lucas.kahlert@cloudandheat.com>