Change delay for retrying connection on Reflector
See discussion on MR !283

The retry mechanism on the Reflector to connect to the API again in case
of disruptions was using an optional max_retry attribute. In case of
disruption, if the connection was possible again, the retry count was
not reset.

As the retry concerns the list and watch mechanism that blocks, we have
no easy mechansim to check if a connection was possible again.

Instead, as the disconnections are to be expected, the max_retry is
simply removed. Instead, the Reflector will try to connect with more and
more delay between the attempts, reaching a maximum after a while, to
prevent reaching too much retry time.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>