Fix conflict of several contexts with same cluster
See issue #288

When a cluster is registered using rok, the different existing contexts
are all used to create a new cluster, one cluster per context. If
different contexts used the same cluster, there was a conflict on the
API, because the name of the cluster was used. However, clusters must
have unique names on the database.

This has now been fixed: the clusters are registered with the context
name, which allows the use of the same cluster several time, as only the
contexts have to be unique.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>