Remove default roles and role bindings from config
See issue #262

Some roles and matching role bindings were previously read from
configuration and taken into consideration as extra roles and role
bindings. This means some parts of internal mechanism were accessible
directly from the configuration files.

As this behavior is not optimal, these defaults objects are removed from
the configuration, and are not used in the API anymore as "default". It
means they now need to be added to the database if RBAC is used.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>