Fix certificate names in API TLS configuration
The name of the configuration parameter to get the path of the
certificate and the key for the API used the prefix "client_", to use
the same as the TLS configuration of the Controllers.

As the certificates are on the server side and are not client, these
parameters have been renamed to match their real use.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>