Add generation of local kubeconfig file with ansible
After using the minikube_cluster.yml and the krake.yml playbook, a
kubeconfig file is created on the Krake VM, for the VM to access easily
the minikube cluster created.

This behavior has been copied to generate also a kubeconfig file on the
machine that started the playbook. This allows the machine to connect
easily to the minikube cluster, for testing purposes for instance. It
allows also the use of the minikube_cluster.yml playbook to create new
minikube clusters easily.

Signed-off-by: Jean Chorin <jean.chorin@cloudandheat.com>